Deploy Trust (#85)
Write ancillary trust file and signal fapolicyd to reload.

Steps applied during deployment

    back up trust file to user data dir
    write ancillary state to the fapolicyd.trust file
    signal fapolicyd

Required a bit of additional configuration work to sketch in the application config. The app config consists of user specific settings and paths. For example in this PR is the users application data directory where we will store backup snapshots of the trust file.

This is not currently taking a backup of the fapolicyd lmdb directory, only the ancillary trust file. It may be determined that the fapolicyd db needs a snapshot as well, but for now there is no need for that.

closes #41
closes #42