Improve app version generation
Adds a pep440 compliant version capture to the build. Sets this as the package version and and exposes it to the application by setting the module version number on the fapolicy_analyzer module.

Sets the About dialog version field to the module version.

Uses a modified CC0 installation of versioneer to provide the version capture.

Note that this is a separate concern from the RPM version calculation, which has a different schema.

Closes #178