Added a refresh button to the Policy Event Analysis Tool (#379)
Added the capability to add tool specific buttons to the toolbar depending on which application tool the user is currently using. Then adds a Refresh button to the Policy Event Analysis tool that will reload either the previously selected audit file or the syslog file. When it refreshes it attempts to reselect the user's drilldown data, if that selected user, subjects, and objects still exist in the reloaded file.