Rule Info API (#417)
A mechanism to provide metadata for a rule.

Metadata can be generated from various sources, for example during the parse that creates a rule from text there can be warnings or errors. Another producer could be a post parse linting process that could generate warnings or suggestions.  This change allows the Rule API to pass any of these types of info along.

This also provides a bool field which indicates whether the rule should be considered valid according to the current configuration of the application.

Closes #390