Rule origin (#418)
Track the file a rule originates in.

This is aimed at supporting the rules.d structure parsing.  It allows the GUI to display the various source files that contribute to the full rule set, while also providing the indication on edit of where the rule should be written out to.