Rule validation wiring
This completes a required step in improving rule parse feedback by connecting the parse results through to the bindings for both success and failure cases.  The improvement here is that on parse failure the parsed text is conveyed through the bindings as a rule that is flagged as invalid, where previously the rule was simply evicted from the DB without any indication there was a failure.

Entries in the rule database are now a choice between a valid and invalid rule state.  Eventually checks will be added to identify the DB itself as valid or invalid, and that will then in turn inform the system as to whether it is valid or invalid.