Reverted the use of python 3.9 ThreadPoolExecutor arguments
Removed the use of the cancel_futures parameter to the ThreadPoolExecutor because it is not available in python 3.6. I've tested this and I think we are good with turning off this cancel functionality. I think it was an issue when you closed the application before the System Trust data loaded completely. The process would hang until the thread pulling the trust data completed, but this isn't an issue now because the rust layer is pre-caching this data during startup. We should be good to revert to the older python 3.6 version now.