Added the ability to remove deleted files from the Ancillary Trust Database
Added the handling of files in Database Admin tool that are in the ancillary trust database but no longer exist on disk. The users can now select these file entries in the tool and choose Untrust. They will be given a warning that the file no longer exists on disk. When they confirm a delete change set for that file will be created.