Fixed issue with Policy Analyzer right click context menu displaying
Fixed the issue where the user had to right click a row 2 times in either the subject or object tables in order to have the context menu display. The first click would select the row and the second would display the context menu. Now a single right click will both select the row and display the menu.