Deserialize relative paths in markers (#528)
Markers were previously being serialized with paths relative to the fapolicyd configuration, however the deserialization machinery expected absolute paths.

This updates the deserialization to handle relative paths, but not enforce that they are relative. A future change will expect relative path in the marker, producing a parse error if a marker uses a directory component in the path.