Write rules.d structure (#533)
Write the rules db to disk while separating files based on origin, placing the files under the configured rules.d directory. This is not required for functionality, as the compiled.rules file is already being written. This instead being done to keep a compliant fapolicyd configuration to ensure the standard fapolicyd workflows will continue to work.

A side effect of this is that the preexisting rules files under rules.d will be deleted prior to writing the structure from the current rules db.. The original rules files are captured in the predeploy snapshot and can be reverted if needed.

Closes #532