Rule change diff (#539)
Provide a diff formatted summary of changes between two rule databases. Taking the approach of diffing the text as it appears in the editor. This provides diffs in our application context, including file markers, and avoids rule number diffs.

This commit adds a function to the bindings with the signature of rules_difference(System, System) which allows you to pass in two systems and receive a string similar to the System#rules_text(). The diff adds prefixes of + and - to all changed lines.

Closes #538