Full daemon reload (#548)
The reload for trust used a signal through a fifo pipe to have fapolicyd reload the trust db. To support rule deployments we need to reload the systemd service. In all cases we will no bring down the daemon to update configuration as there is no way we can guarantee that the write to disk would be successful while fapolicyd is up.

Closes #544
Closes #547