Change post deployment revert so changes are kept in the UI (#609)
Changed the UI behavior so changes are kept in the UI if the user decides to rollback the deploy when the post-deployment revert dialog. The functionality of loading a saved session file should be the same. When loading the file, current changes are cleared and the set of changes from the session file are loaded.

Closes #550