Prevent empty rules (#623)
Detect empty rules and throw a parse error to prevent the system from deploying and ultimately blowing up the fapolicyd reload.

Classifying the empty-rule failure in same bucket as a malformed-marker parse, which means it stops the changeset from being populated.  This seems to be the right spot in the flow to put a halt to things.

This impacts the rule changeset bindings by dropping the `set` call which never throws, and forcing the use of the `parse` call that may throw.  Forcing the client to be aware of the potential parse error here is important.

Closes #604


Co-authored-by: Shawn Dorsch <dorschs@ctc.com>