Improved RPM build (#649)
The goal of this work is to resolve the build-time dependency issues
encountered with both Rust crates and Python packages.

To address the dependency issues we have implemented a vendoring process
for Rust that vendors the crates that have no official packages and
lists the available crates as `BuildRequires`. On the Python side we
have several items missing in RHEL builds which we address by listing
SourceX dependencies from PyPI and installing those in the `%prep`
stage.

This process is tested end-to-end by generating a SRPM which is then
used to create a RPM in an offline build. The GitHub CI process
implements this and there is also a Fedor Copr build that also proves
out the approach.

The missing pieces that we vendor are what we are referencing when
saying _there is a gap in the build-time dependencies_.

This PR resolves the build time dependency issues, and a few other spec
and rpm build related issues along the way. The process defined here
accomplishes that by vendoring the Python and Rust sources that are
without official support into the SRPM. The SRPM is then be used to
create both Fedora and RHEL 8 RPMs which use official packages where
possible and vendored packaged where not.

There are no run-time dependencies being vendored.

## Rationale

The following documents the what and why of this commit.

### The What

1. Adds spec files that support Fedora and EL which supports offline
builds using vendored Python and Rust
2. Add scripts to vendor all or a subset of the Rust crate dependency,
the remainder are added as SourceX
3. Add scripts to determine what crates need vendored based on the
availability of official packages
4. Add SourceX dependencies for Python build tools unavailable on RHEL
5. Add Makefile to configure Fedora Copr build environment
6. Add GitHub CI workflow to build SRPM, and then rebuild a RPM from it
7. Add matrix to the SRPM workflow to run on fc:37, rawhide, and UBI 8
8. Capture SRPM build artifacts in the releases, including spec file,
srpm, rpm, vendor tarballs
9. Rewrite the spec file captured in release to point to the url of the
release artifact rather than the local path
10. Replace the original RPM build with the new SRPM build

### The Why

1. There are build time dependencies missing for Rust in both Fedora and
RHEL
2. There are build time dependencies missing for Python in RHEL
3. We must be able to build offline, using vendored dependencies, from a
SRPM
4. We need to build for RHEL 8 for distribution, but also want to
prepare for next-gen features from Fedora Rawhide

### Other Whats
1. Container file SRPM builds for RHEL and Rawhide
2. Vagrant Rawhide configuration
3. Playbook will provision from Copr when version is `latest` (default
if unspecified)
4. Playbook will provision from GitHub when version is `latest-release`
or a specific release (eg. `v0.6.0`)

### Why are there two specs?

The "production" spec now lives under the project root, this is
currently FC only. The scripts/srpm spec is the next version with RHEL
support. Keeping both to ensure we don't have to deal with any drift in
the future when RHEL moves into the production spec.

Closes #506
Closes #622
Closes #627
Closes #637 
Closes #641