Trust callbacks (#691)
Async loading of trust by implementing Python callbacks within the
bindings to allow notification of progress during the trust loading
process.

Alongside the callback mechanism this adds the ability to merge updated
trust entries into the System without the need for a change operation.
This allows trust that has been calculated externally to be added to the
System, post-init.

With these enhancements in place the default System can defer the
"actual" checking, only loading the trust sources. This greatly speeds
up the load time, and removed the need for costly hashing and disk io at
startup time. Deferred loading will eventually be the default, however
for the time being to avoid breaking master we will provide a binding
function to load an unchecked system, `unchecked_system`. This will be
removed and the System ctor will use unchecked after the UI has
integrated the callback behavior.

The "actual" checking has a new binding function `check_disk_trust` that
spawns a number of threads to traverse the trust sources and hash the
files on disk. This call accepts two Python function pointers as
callbacks: (1) call when a new batch of data is ready, which is passed
as a list of Trust entries, (2) call when all data processing has
completed.

The callback pattern appears to be reliable and supports concurrency on
the backend. There is great potential here for integration with the
Redux machinery. There is a minimal example of this in the
example/async_trust.py example, which is also a good entrypoint to see
the overall functionality in this PR in action.

Closes #690