Fix locale install (#733)
The application locales were not being included in the RPM. This PR
rearranges the location of locales, bringing them up to a top level
directory, and fixes up the installation process.

Adds additional content to the man page and fixes an issue installing it
#200.

This also fixes up a few annoyances caused from the vendoring refactor
in #730.

Also adds a check stage to CI that performs a RPM install and displays
the rpmdb metadata for it.

More #715
Closes #720
# Conflicts:
#	.copr/Makefile
#	Makefile
#	fapolicy-analyzer.spec
#	scripts/srpm/fapolicy-analyzer.spec