Distinct trust checks (#755)
Improve #691 by adding ability to distinguish between type of trust
being checked.

Rather than add a flag to the type, or separate return batches, this
adds separate entrypoints. The callback handlers will be responsible for
knowing which type they are receiving.

This also changes the return value of the `check_*` call to represent
the total number of records to be checked. The count parameter of the
update callback is now the total number checked as of that callback,
such that the `check_*` return value should equal the callback count
value on the final batch.

See the updated reference impl in `examples/async_trust.py`

#690