Access path to app config through bindings
The Python side may need access to the configuration file.  While there is no need to provide fine grained config modeling in the bindings, the ability to have a single source pointing to the file is useful.

This adds a single function `config_file_path` that provides the path to the application config as a string.

Closes #786