Bindings for app config path (#788)
Provides the path to the app config through bindings

The Python side needs access to the app configuration, and while
there is no need to provide fine grained config modeling in the
bindings, having a single authoritative path is useful.

This adds a function `config_file_path` that provides the path to
the application config toml as a string.

This also changes the config file name from `fapolicy-analyzer.toml` to
`config.toml`. The file is already namespaced by the config dir, there
is no need to duplicate it in the filename. This improves the
readability of documentation and other discussions.

Closes #786