Profiler fixes (#809)
Address a couple of bugs found in profiler after #770

Fixes a bug in the profiler reducer that only happened when there was
env, regressed when `env_dict` was added to the state. This was causing
silent crashes likely because the redux error handling isn't wired the
best way. Will revisit this.

Fixes a bug on RHEL 8 where the fapolicyd output was not being written
to file due to SELinux denial. After extensive trials the best approach
was to drop systemd control and implement our own daemon control.
Another benefit is that this avoids a couple bugs on the el8 systemd
v239 that had to be worked around. This also avoids the somewhat suspect
practice of writing the unit file on demand, though in hindsight we
could of used systemd-run.

Add fapolicyd runtime log parsing that allows the profiler to wait until
the fapolicyd service is ready before kicking off the target process.
Even with the service integration mentioned in #531 the timing issues
were not resolved. This is because the "up" status of the daemon does
not mean "ready to observe events", there is a space between the two the
duration of which depends on the size of the trust database. We were
aware of this previously, it just got kicked down the road until now.

Closes #531
Closes #808