Customizable initial view (#789)
Select the initial application view through the config.toml.

Right now this is manual process. To set your initial view update your 
config.toml file with a section like

```toml
[ui]
initial_view = 'trust'
```

The possible values are "trust", "rules", "debug log", "syslog", or
"profiler". This is an optional setting, and the default value when
not specified is "rules".

As part of this change the I removed the disabling of the trust menu
option when you are on the trust screen. If the user is on the trust
page and they hit the trust menu option again it just refreshes the
page.

Closes #773 
Closes #804