Build updates for epel8 (#794)
Changes applied to the Fedora repo prior to merging for EPEL 8.

The most significant change is to remove the parallel support for Fedora
in the el8 spec file. The el8 version is noisy and odd due to the Python
patching, keeping them together pollutes all the other versions.
Starting at el9 we can start to integrate it into the main spec file.

The pytz dependency is now available on pypi and was updated to fetch
with the `pypi_source` macro.

Add recommended Rust compiler flags as recommended in Fedora. Had to do
this manually as the rust macro to do so is not available for el8.

Cleaned up the readme and comments documenting the spec file.

There were some other changes applied in the rpms pr that were not
applied here due to being 0.6 specific.

- https://src.fedoraproject.org/rpms/fapolicy-analyzer/pull-request/3

Updates the rpm matrix to bring the Fedora builds up to date now that 38
has been branched.

Adds a step to tag the doc version on releases. This is to ensure the
doc repo has matching tags to preserve the build-to-doc relationship.
Only required on releases, other builds can specify manually or float.

Closes #757
Closes #793