Fix ancillary trust loading when swapping session files (#840)
Fixes some threading/callback issues when a system revert happens
followed by a set of changesets being applied. This happens when a user
loads a new session file with changesets in it, or the system is
reverted using the post-deployment revert option. In both cases an async
trust check is kicked off from the revert. When the changesets are then
applied a trust check using the new System object was not initiated.

closes #828 
closes #827