Build for epel 9 (#864)
Adds epel 9 builds in CI and for local containers.

The recipe here is a mix between the Fedora and epel 8 builds. There are
quite a few missing epel components yet, they are being requested as
part of #775. Eventually epel 9 will closely resemble the FC build and
the specs should be merged. However for now we will maintain a separate
spec like is done for epel 8.

This also adds a version suffix to the release artifacts. The intent is
to guard against human error when manually moving artifacts between
GitHub and Fedora releases.


Closes #821
Closes #865