Fix spec version patching (#878)
Some previously undetected regressions were brought to light after
merging the implementation of #821 in #865.

The background here is that the version defined in the spec files can
be patched during CI builds for master or a PR branch. This is done to
allow these builds to be versioned with qualifiers that identify them
as development builds.

The spec version patching had a couple issues that were introduced back
when removing the pull-request-target support. This PR fixes those and
some other other issues causing the work from #865 not to succeed on
master.