Fix editor state after deploy (#935)
This fixes an issue where the text editors were erroneously notifying of
existing changes after a deployment.

For reference see the changes in #727 and #921. The text clearing in the
former is clashing with the flagging in the latter. This PR addressed it
by flipping a flag right after the point where the text is cleared.

Closes #934