Check crate semver on pull request (#960)
Adds a semver check on pull requests

The goal is to check compat issues before hitting master. The resolution
to a failure here is usually to bump the version in the offending
Cargo.toml, with the alternative fix being to alter the offending code.

This sets `continue-on-error: true` to avoid disturbing the rest of the
workflow. Ultimately it is up to the committer to accept or reject the
offending patch.

Closes #704