Reload rules while profiling (#990)
Allow rules to be loaded dynamically into a profiling session.

This allows for a better test / update cycle while profiling as the
profiler does not have to be shutdown to update rules.

This also adds a new concept, the "Rule Identity", which provides a
sha256 hash of the rule database to allow change monitoring across both
profiler and systemd daemon execution. This hash is based on the
compiled rules to provide a precise content based hash.

Closes #985
Closes #989
# Conflicts:
#	crates/pyo3/Cargo.toml