Bump flit-core back on el9 (#1035)
This is an update that was applied in the Fedora repo to resolve an
issue publishing the 1.4.0 rpm.

The flit-core update to 3.9.0 pulled in a tomli dependency that was too
new for the build system. The difference between CI and Koji was the
inability to pull the PyPi dependencies brought tomli issue to light.