Fix trust init error (#1038)
Handle leading spaces on sha1 trust hash

The trust init was failing on a rsplit on spaces when the trust fields
were separated by more than a single space. Update that to break the
parse into multiple split operations that can be trimmed individually.

The debugging process for this made use of Copr, and while doing so it
became clear that the number scheme for PR builds needed improved. This
updates the format to build 0.PR.COMMIT, which provides a better flow
for these temporary versions than the previous 0.0.PR which would
require overwriting the same version on every commit and also did not
provide any indication in the build where the build originated.

Closes #1036