Stats view and backend (#1048)
Implements the fapolicyd cache stats view.

Adds a backend for parsing and tracking stats taken from the
fapolicyd.state file. The backend watches for writes to the file and
captures the state into an in-memory database.

The GUI implements a simple view that displays the current state file
and a graph of object cache performance. The graph is populated by a
time-limited historical view of the state file. This is provided by the
backend and linked into the stats view redux feature.

Additional graphs may be added to the cache view in the future. The
stats data may also be displayed on the profiler view in the future.

The stats backend can be leveraged to display stats from the CLI as well
as the core metrics tracking is done on the backend in rust.

Closes #621